feat(#10) - Persistence Test MyBatis , Persistance -> Persistence 오타 수정
