+++ src/main/java/com/flab/infrun/lecture/infrastructure/persistence/jdbcTemplate/LectureJdbcTemplateRepository.java
@@ -1,6 +1,6 @@
-package com.flab.infrun.lecture.infrastructure.persistance.jdbcTemplate;
+package com.flab.infrun.lecture.infrastructure.persistence.jdbcTemplate;
 
-import com.flab.infrun.lecture.domain.LectureVideoFile;
+import com.flab.infrun.lecture.domain.Lecture;
 import java.util.Map;
 import java.util.Optional;
 import javax.sql.DataSource;
@@ -9,38 +9,42 @@ import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
 import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 
-public class LectureVideoFileJdbcTemplateRepository {
+public class LectureJdbcTemplateRepository {
 
     private final NamedParameterJdbcTemplate jdbcTemplate;
     private final SimpleJdbcInsert jdbcInsert;
 
-    public LectureVideoFileJdbcTemplateRepository(DataSource dataSource) {
+    public LectureJdbcTemplateRepository(DataSource dataSource) {
         this.jdbcTemplate = new NamedParameterJdbcTemplate(dataSource);
         this.jdbcInsert = new SimpleJdbcInsert(dataSource)
-            .withTableName("file")
+            .withTableName("lecture")
             .usingGeneratedKeyColumns("id");
+
     }
 
-    public Optional<LectureVideoFile> findById(Long id) {
-        String sql = "select id , name, url FROM file where id = :id";
+    public Optional<Lecture> findById(Long id) {
+        String sql = "select id , name, price, introduce FROM LECTURE where id = :id";
         Map<String, Object> param = Map.of("id", id);
-        LectureVideoFile lecture = jdbcTemplate.queryForObject(sql, param, itemRowMapper());
-
+        Lecture lecture = jdbcTemplate.queryForObject(sql, param, itemRowMapper());
         return Optional.ofNullable(lecture);
     }
 
-    private RowMapper<LectureVideoFile> itemRowMapper() {
+    private RowMapper<Lecture> itemRowMapper() {
         return (rs, rowNum) -> {
-            LectureVideoFile lecture = LectureVideoFile.of(
-                rs.getString("url"),
-                rs.getString("name")
+            Lecture lecture = Lecture.of(
+                rs.getString("name"),
+                rs.getInt("price"),
+                rs.getInt("level"),
+                rs.getString("skill"),
+                rs.getString("introduce"),
+                rs.getLong("userId")
             );
             lecture.setId(rs.getLong("id"));
             return lecture;
         };
     }
 
-    public LectureVideoFile save(LectureVideoFile entity) {
+    public Lecture save(Lecture entity) {
 
         BeanPropertySqlParameterSource param = new BeanPropertySqlParameterSource(
             entity);
