+++ src/main/java/com/flab/infrun/lecture/infrastructure/persistence/jdbcTemplate/LectureVideoFileJdbcTemplateRepository.java
@@ -1,6 +1,6 @@
-package com.flab.infrun.lecture.infrastructure.persistance.jdbcTemplate;
+package com.flab.infrun.lecture.infrastructure.persistence.jdbcTemplate;
 
-import com.flab.infrun.lecture.domain.Lecture;
+import com.flab.infrun.lecture.domain.LectureVideoFile;
 import java.util.Map;
 import java.util.Optional;
 import javax.sql.DataSource;
@@ -9,42 +9,38 @@ import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
 import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 
-public class LectureJdbcTemplateRepository {
+public class LectureVideoFileJdbcTemplateRepository {
 
     private final NamedParameterJdbcTemplate jdbcTemplate;
     private final SimpleJdbcInsert jdbcInsert;
 
-    public LectureJdbcTemplateRepository(DataSource dataSource) {
+    public LectureVideoFileJdbcTemplateRepository(DataSource dataSource) {
         this.jdbcTemplate = new NamedParameterJdbcTemplate(dataSource);
         this.jdbcInsert = new SimpleJdbcInsert(dataSource)
-            .withTableName("lecture")
+            .withTableName("file")
             .usingGeneratedKeyColumns("id");
-
     }
 
-    public Optional<Lecture> findById(Long id) {
-        String sql = "select id , name, price, introduce FROM LECTURE where id = :id";
+    public Optional<LectureVideoFile> findById(Long id) {
+        String sql = "select id , name, url FROM file where id = :id";
         Map<String, Object> param = Map.of("id", id);
-        Lecture lecture = jdbcTemplate.queryForObject(sql, param, itemRowMapper());
+        LectureVideoFile lecture = jdbcTemplate.queryForObject(sql, param, itemRowMapper());
+
         return Optional.ofNullable(lecture);
     }
 
-    private RowMapper<Lecture> itemRowMapper() {
+    private RowMapper<LectureVideoFile> itemRowMapper() {
         return (rs, rowNum) -> {
-            Lecture lecture = Lecture.of(
-                rs.getString("name"),
-                rs.getInt("price"),
-                rs.getInt("level"),
-                rs.getString("skill"),
-                rs.getString("introduce"),
-                rs.getLong("userId")
+            LectureVideoFile lecture = LectureVideoFile.of(
+                rs.getString("url"),
+                rs.getString("name")
             );
             lecture.setId(rs.getLong("id"));
             return lecture;
         };
     }
 
-    public Lecture save(Lecture entity) {
+    public LectureVideoFile save(LectureVideoFile entity) {
 
         BeanPropertySqlParameterSource param = new BeanPropertySqlParameterSource(
             entity);
