+++ src/main/resources/com/flab/infrun/lecture/infrastructure/persistence/mybatis/mapper/LectureDetailMyBatisMapper.xml
@@ -2,7 +2,7 @@
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Config 3.0//EN"
   "https://mybatis.org/dtd/mybatis-3-mapper.dtd">
 <mapper
-  namespace="com.flab.infrun.lecture.infrastructure.persistance.mybatis.mapper.LectureDetailMyBatisMapper">
+  namespace="com.flab.infrun.lecture.infrastructure.persistence.mybatis.mapper.LectureDetailMyBatisMapper">
 
   <select id="findById" resultType="com.flab.infrun.lecture.domain.LectureDetail">
     SELECT ID
