+++ src/main/resources/com/flab/infrun/lecture/infrastructure/persistence/mybatis/mapper/LectureMyBatisMapper.xml
@@ -2,7 +2,7 @@
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Config 3.0//EN"
   "https://mybatis.org/dtd/mybatis-3-mapper.dtd">
 <mapper
-  namespace="com.flab.infrun.lecture.infrastructure.persistance.mybatis.mapper.LectureMyBatisMapper">
+  namespace="com.flab.infrun.lecture.infrastructure.persistence.mybatis.mapper.LectureMyBatisMapper">
 
   <select id="findById" resultType="com.flab.infrun.lecture.domain.Lecture">
     SELECT id, name, nvl(price,0), nvl(lecture_level,0), skill, introduce, nvl(user_id,0)
