+++ src/main/resources/com/flab/infrun/lecture/infrastructure/persistence/mybatis/mapper/LectureVideoFileMyBatisMapper.xml
@@ -2,7 +2,7 @@
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Config 3.0//EN"
   "https://mybatis.org/dtd/mybatis-3-mapper.dtd">
 <mapper
-  namespace="com.flab.infrun.lecture.infrastructure.persistance.mybatis.mapper.LectureVideoFileMyBatisMapper">
+  namespace="com.flab.infrun.lecture.infrastructure.persistence.mybatis.mapper.LectureVideoFileMyBatisMapper">
 
   <select id="findById" resultType="com.flab.infrun.lecture.domain.LectureVideoFile">
     SELECT ID , URL, NAME
