+++ src/test/java/com/flab/infrun/lecture/infrastructure/persistence/LectureDetailRepositoryTest.java
@@ -1,20 +1,20 @@
-package com.flab.infrun.lecture.infrastructure.persistance;
+package com.flab.infrun.lecture.infrastructure.persistence;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import com.flab.infrun.lecture.domain.LectureDetail;
-import com.flab.infrun.lecture.domain.LectureDetailRepository;
+import com.flab.infrun.lecture.infrastructure.persistence.mybatis.mapper.LectureDetailMyBatisMapper;
 import java.util.Optional;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import org.mybatis.spring.boot.test.autoconfigure.MybatisTest;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
 
-@SpringBootTest
+@MybatisTest
 class LectureDetailRepositoryTest {
 
     @Autowired
-    private LectureDetailRepository adapter;
+    private LectureDetailMyBatisMapper mapper;
 
     //todo-Transaction으로 처리
     @Test
@@ -22,8 +22,8 @@ class LectureDetailRepositoryTest {
     void save() {
         LectureDetail lectureDetail = LectureDetail.of("1", "강의 A의 챕터1", 1L, 2L);
 
-        LectureDetail saved = adapter.save(lectureDetail);
-        Optional<LectureDetail> lecture1 = adapter.findById(saved.getId());
+        int saved = mapper.save(lectureDetail);
+        Optional<LectureDetail> lecture1 = mapper.findById((long) saved);
 
         assertThat(lecture1.get()).isEqualTo(lectureDetail);
     }
