+++ src/test/java/com/flab/infrun/lecture/infrastructure/persistence/LectureRepositoryTest.java
@@ -1,19 +1,20 @@
-package com.flab.infrun.lecture.infrastructure.persistance;
+package com.flab.infrun.lecture.infrastructure.persistence;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import com.flab.infrun.lecture.domain.Lecture;
+import com.flab.infrun.lecture.infrastructure.persistence.mybatis.mapper.LectureMyBatisMapper;
 import java.util.Optional;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import org.mybatis.spring.boot.test.autoconfigure.MybatisTest;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
 
-@SpringBootTest
+@MybatisTest
 class LectureRepositoryTest {
 
     @Autowired
-    private LectureRepositoryAdapter adapter;
+    private LectureMyBatisMapper mapper;
 
     //todo-Transaction으로 처리
     @Test
@@ -22,8 +23,8 @@ class LectureRepositoryTest {
         Lecture lecture = Lecture.of("Lecture1", 20000, 1, "JAVA", "이것은 강의 A",
             1L);
 
-        Lecture saved = adapter.save(lecture);
-        Optional<Lecture> lecture1 = adapter.findById(saved.getId());
+        int saved = mapper.save(lecture);
+        Optional<Lecture> lecture1 = mapper.findById((long) saved);
 
         assertThat(lecture1.get()).isEqualTo(lecture);
     }
