+++ src/test/java/com/flab/infrun/lecture/infrastructure/persistence/LectureVideoFileRepositoryTest.java
@@ -1,20 +1,20 @@
-package com.flab.infrun.lecture.infrastructure.persistance;
+package com.flab.infrun.lecture.infrastructure.persistence;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import com.flab.infrun.lecture.domain.LectureVideoFile;
-import com.flab.infrun.lecture.domain.LectureVideoFileRepository;
+import com.flab.infrun.lecture.infrastructure.persistence.mybatis.mapper.LectureVideoFileMyBatisMapper;
 import java.util.Optional;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import org.mybatis.spring.boot.test.autoconfigure.MybatisTest;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
 
-@SpringBootTest
+@MybatisTest
 class LectureVideoFileRepositoryTest {
 
     @Autowired
-    private LectureVideoFileRepository adapter;
+    private LectureVideoFileMyBatisMapper mapper;
 
     //todo-Transaction으로 처리
     @Test
@@ -22,9 +22,9 @@ class LectureVideoFileRepositoryTest {
     void save() {
         LectureVideoFile lectureVideoFile = LectureVideoFile.of("/lecture", "파일명1.pdf");
 
-        LectureVideoFile saved = adapter.save(lectureVideoFile);
+        int saved = mapper.save(lectureVideoFile);
 
-        Optional<LectureVideoFile> lecture1 = adapter.findById(saved.getId());
+        Optional<LectureVideoFile> lecture1 = mapper.findById((long) saved);
         assertThat(lecture1.get()).isEqualTo(lectureVideoFile);
     }
 }
